mvcc: refactor RowID.row_id to be either i64 or a record
to make interoperability with VDBE opcodes easier, we are going
to store versioned index rows (unlike in Hekaton). for this reason
refactor RowID.row_id to either be an i64 or an index key (record).

this lets us store a SkipMap from index key to a list of index row
versions like normal table row versions.