fix/sim: abandon the entire property if precondition fails
e.g. in "ReadYourUpdatesBack", we first do an UPDATE and then a
SELECT to verify that those updates were actually made.

Problem is the SELECT assertion was made even if the UPDATE failed.

Fix: change interaction handling so that if a precondition in a
property fails, the entire property is abandoned.