fix: update schema if DDL commit succeeded, checkpoint failed
This is already done in abort() but if the checkpoint error is
returned directly from commit_tx() instead of in an IO completion,
then the Tx state will be cleared and the DDL change won't be applied.
So, do it explicitly in step_end_write_txn() too.

Fixes this failing seed:

cargo run -- --memory-io --seed 12902559987470385066 --disable-heuristic-shrinking